Read HTP data

Read HTP data

Usage

read_HTP(data, genotype, time, plot, traits, row, range)

Arguments

data

A data.frame in a wide format.

genotype

A character string indicating the column in data that contains genotypes.

time

A character string indicating the column in data that contains the time points.

plot

A character string indicating the column in data that contains the plot IDs.

traits

A character vector specifying the traits for which the models should be fitted.

row

A character string indicating the column in data that contains the row coordinates.

range

A character string indicating the column in data that contains the range coordinates.

Value

An object of class read_HTP, with a list of:

summ_traits

A data.frame containing a summary of the traits.

exp_design_resum

A data.frame containing a summary of the experimental design.

locals_min_max

A data.frame containing local maximums and minimuns

dt_long

A data.frame in long format.

Examples

library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
print(results)
#> $summ_traits
#> # A tibble: 16 × 12
#>    trait   time     Min    Mean   Median      Max       SD       CV     n  miss
#>    <chr>  <dbl>   <dbl>   <dbl>    <dbl>    <dbl>    <dbl>    <dbl> <int> <int>
#>  1 Canopy    29  0       0        0        0       0       NaN        196     0
#>  2 Canopy    36  0       2.86     1.77    14.5     3.12      1.09     196     0
#>  3 Canopy    42  0.731  22.7     22.3     44.1     9.00      0.397    196     0
#>  4 Canopy    56 32.8    74.0     73.7     99.2    12.2       0.165    196     0
#>  5 Canopy    76 89.6    99.7    100      100       1.05      0.0105   196     0
#>  6 Canopy    92 63.1    97.1     98.5    100.      4.24      0.0436   196     0
#>  7 Canopy   100 27.6    82.7     87.5     99.7    14.8       0.179    196     0
#>  8 Canopy   108  5.15   68.8     74.4     99.1    21.0       0.305    196     0
#>  9 PH        29 NA      NA       NA       NA      NA        NA          0   196
#> 10 PH        36 -0.0107  0.0232   0.0236   0.0513  0.00877   0.377    176    20
#> 11 PH        42 -0.0350  0.0719   0.0730   0.205   0.0452    0.629    196     0
#> 12 PH        56 -0.122   0.0353   0.0218   0.346   0.0802    2.27     196     0
#> 13 PH        76  0.352   0.563    0.564    0.756   0.0759    0.135    196     0
#> 14 PH        92  0.220   0.359    0.353    0.527   0.0557    0.155    196     0
#> 15 PH       100  0.0510  0.234    0.235    0.413   0.0765    0.327    196     0
#> 16 PH       108  0.261   0.724    0.757    0.988   0.170     0.235    196     0
#> # ℹ 2 more variables: `miss%` <dbl>, `neg%` <dbl>
#> 
#> $exp_design_resum
#> # A tibble: 1 × 6
#>       n n_gen n_row n_range num_of_reps num_of_gen
#>   <int> <int> <int>   <int> <chr>       <chr>     
#> 1   196   185    14      14 1_2         174_11    
#> 
#> $locals_min_max
#> # A tibble: 2 × 3
#>   trait  local_min_at local_max_at
#>   <chr>  <chr>        <chr>       
#> 1 Canopy ""           76          
#> 2 PH     "56_100"     42_76       
#> 
#> $dt_long
#> # A tibble: 3,136 × 7
#>     time  plot   row range genotype  trait    value
#>    <dbl> <dbl> <dbl> <dbl> <chr>     <chr>    <dbl>
#>  1    29     1     1     1 W17037-24 Canopy  0     
#>  2    29     1     1     1 W17037-24 PH     NA     
#>  3    36     1     1     1 W17037-24 Canopy  0.648 
#>  4    36     1     1     1 W17037-24 PH      0.0154
#>  5    42     1     1     1 W17037-24 Canopy 14.4   
#>  6    42     1     1     1 W17037-24 PH      0.0173
#>  7    56     1     1     1 W17037-24 Canopy 74.5   
#>  8    56     1     1     1 W17037-24 PH      0.0941
#>  9    76     1     1     1 W17037-24 Canopy 99.7   
#> 10    76     1     1     1 W17037-24 PH      0.456 
#> # ℹ 3,126 more rows
#> 
#> attr(,"class")
#> [1] "read_HTP"